requirements.txt
apispec==6.6.1
astroid==3.0.3
bcrypt==4.1.2
black==24.4.2
blinker==1.7.0
cachelib==0.13.0
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.0
execnet==2.0.2
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Mpesa==1.2.8
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flask-vite==0.5.0
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mccabe==0.7.0
msgspec==0.18.6
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==23.2
patchelf==0.17.2.1
pathspec==0.12.1
phonenumbers==8.13.39
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
PyJWT==2.8.0
pylint==3.0.3
pyparsing==3.1.2
pyproject-api==1.6.1
pytest==8.0.0
pytest-sugar==1.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.25
temp_email_validator==0.1
termcolor==2.4.0
tomli==2.0.1
tomlkit==0.12.3
tox==4.12.1
trafaret==2.0.2
types-Flask-Cors==4.0.0.20240405
typing_extensions==4.9.0
urllib3==2.2.0
uuid==1.30
virtualenv==20.25.0
waitress==3.0.0
webargs==8.4.0
Werkzeug==3.0.1
WTForms==3.1.2
